优化：从ifox-common中分离spring data jpa的依赖，独立module : ifox-jpa
