+++ email-service/src/main/resources/application.yml
@@ -30,17 +30,3 @@ spring:
   http.multipart:
       max-request-size: 30MB
       max-file-size: 30MB
-  datasource:
-    driverClassName: com.mysql.jdbc.Driver
-    username: root
-    password: Aswd1234
-    url: jdbc:mysql://localhost:3306/spring_data_test
-  jpa:
-    database: MYSQL
-    hibernate:
-      ddl-auto: update
-      naming-strategy: org.hibernate.cfg.ImprovedNamingStrategy
-    properties:
-      hibernate:
-        dialect: org.hibernate.dialect.MySQL5Dialect
-    show-sql: false
