+++ ifox-common/src/main/java/com/ifox/platform/common/page/SimplePage.java
@@ -1,7 +1,6 @@
 package com.ifox.platform.common.page;
 
 import com.ifox.platform.common.rest.response.PageResponseDetail;
-import org.springframework.data.domain.Page;
 
 import java.util.List;
 
@@ -203,15 +202,6 @@ public class SimplePage<T> implements Pageable {
 
     public void setContent(List<T> content) {
         this.content = content;
-    }
-
-    /**
-     * 通过Page对象初始化
-     * @param page Page
-     * @return SimplePage
-     */
-    public SimplePage<T> initWithSpringDataPage(Page<T> page) {
-        return new SimplePage<T>(page.getNumber(), page.getSize(), (int)page.getTotalElements(), page.getContent());
     }
 
     @Override
