+++ ifox-jpa/src/main/java/com/ifox/platform/jpa/converter/PageRequestConverter.java
@@ -1,0 +1,17 @@
+package com.ifox.platform.jpa.converter;
+
+import com.ifox.platform.common.rest.request.PageRequest;
+import org.springframework.data.domain.Pageable;
+
+public class PageRequestConverter {
+
+    /**
+     * 转换PageRequest为spring data Pageable
+     * @param pageRequest PageRequest
+     * @return Pageable
+     */
+    public static Pageable convertToSpringDataPageable(PageRequest pageRequest) {
+        return new org.springframework.data.domain.PageRequest(pageRequest.getPageNo(), pageRequest.getPageSize(), SimpleOrderConverter.convertToSort(pageRequest.getSimpleOrderList()));
+    }
+
+}
