+++ system-service/build.gradle
@@ -12,8 +12,8 @@ repositories {
 }
 
 dependencies {
-    compile project(":ifox-common")
     compile project(":ifox-utility")
+    compile project(":ifox-jpa")
 
     testCompile group: 'junit', name: 'junit', version: '4.12'
 }
