+++ system-service/src/main/java/com/ifox/platform/system/service/AdminUserService.java
@@ -2,7 +2,6 @@ package com.ifox.platform.system.service;
 
 import com.ifox.platform.common.exception.BuildinSystemException;
 import com.ifox.platform.common.page.SimplePage;
-import com.ifox.platform.system.dto.AdminUserDTO;
 import com.ifox.platform.system.entity.AdminUserEO;
 import com.ifox.platform.system.exception.NotFoundAdminUserException;
 import com.ifox.platform.system.exception.RepeatedAdminUserException;
